requirements.txt
asgiref==3.2.7
autopep8==1.5.2
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.5
django-appconf==1.0.4
django-imagekit==4.0.2
django-import-export==2.0.2
django-widget-tweaks==1.4.8
entrypoints==0.3
et-xmlfile==1.0.1
flake8==3.7.9
jdcal==1.4.1
MarkupPy==1.14
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.0.3
pilkit==2.0
Pillow==7.1.1
psycopg2==2.8.5
pycodestyle==2.5.0
pyflakes==2.1.1
pytz==2019.3
PyYAML==5.3.1
six==1.14.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
tablib==1.1.0
xlrd==1.2.0
xlwt==1.3.0

requests~=2.23.0
djangorestframework~=3.11.0
reportlab~=3.5.42
pdfkit~=0.6.0
pandas~=1.0.5